% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num_as_nor_invlog2_1}
\alias{format_num_as_nor_invlog2_1}
\alias{format_num_as_nor_invlog2_2}
\title{Format numbers on inverse log-2 scale with Norwegian conventions (1 decimal place)}
\usage{
format_num_as_nor_invlog2_1(x)

format_num_as_nor_invlog2_2(x)
}
\arguments{
\item{x}{Numeric vector of values on the log-2 scale.}
}
\value{
Character vector of Norwegian-formatted numbers.

Character vector of Norwegian-formatted numbers.
}
\description{
Applies \code{2^x}, then formats with Norwegian conventions and 1 decimal place.

Applies \code{2^x}, then formats with Norwegian conventions and 2 decimal places.
}
\examples{
# Transform log2 values back to original scale
format_num_as_nor_invlog2_1(c(1, 2, 3))  # 2^1, 2^2, 2^3
format_num_as_nor_invlog2_2(c(1, 2, 3))
}
\seealso{
Other csstyle_formatters:
\code{\link[=format_num_as_nor_invlog10_1]{format_num_as_nor_invlog10_1()}},
\code{\link[=format_num_as_nor_invloge_1]{format_num_as_nor_invloge_1()}},
\code{\link[=format_num_as_nor_num_0]{format_num_as_nor_num_0()}},
\code{\link[=format_num_as_nor_per100k_0]{format_num_as_nor_per100k_0()}},
\code{\link[=format_num_as_nor_perc_0]{format_num_as_nor_perc_0()}}
}
\concept{csstyle_formatters}
