% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num_as_nor_invloge_1}
\alias{format_num_as_nor_invloge_1}
\alias{format_num_as_nor_invloge_2}
\title{Format numbers on inverse natural log scale with Norwegian conventions (1 digit)}
\usage{
format_num_as_nor_invloge_1(x)

format_num_as_nor_invloge_2(x)
}
\arguments{
\item{x}{Numeric vector (natural log scale values)}
}
\value{
Character vector with Norwegian-formatted numbers

Character vector with Norwegian-formatted numbers
}
\description{
Formats values by applying exp(x) transformation then formatting with Norwegian conventions and 1 decimal place.

Formats values by applying exp(x) transformation then formatting with Norwegian conventions and 2 decimal places.
}
\examples{
# Transform natural log values back to original scale
format_num_as_nor_invloge_1(c(0, 1, 2))  # exp(0), exp(1), exp(2)
format_num_as_nor_invloge_2(c(0, 1, 2))
}
\seealso{
Other csstyle_formatters:
\code{\link[=format_num_as_nor_invlog10_1]{format_num_as_nor_invlog10_1()}},
\code{\link[=format_num_as_nor_invlog2_1]{format_num_as_nor_invlog2_1()}},
\code{\link[=format_num_as_nor_num_0]{format_num_as_nor_num_0()}},
\code{\link[=format_num_as_nor_per100k_0]{format_num_as_nor_per100k_0()}},
\code{\link[=format_num_as_nor_perc_0]{format_num_as_nor_perc_0()}}
}
\concept{csstyle_formatters}
