% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{pretty_breaks}
\alias{pretty_breaks}
\title{Pretty breaks}
\usage{
pretty_breaks(n = 5, formatting_fn = csstyle::format_num_as_nor_num_0, ...)
}
\arguments{
\item{n}{Number of desired breaks. Default 5.}

\item{formatting_fn}{Function applied to the break labels. Default
\code{format_num_as_nor_num_0()}.}

\item{...}{Further arguments passed to \code{pretty()}.}
}
\value{
A function that you can use for axis breaks.
}
\description{
Creates a function that generates pretty axis breaks with
  Norwegian number formatting.
}
\examples{
library(ggplot2)

# Use pretty breaks with default formatting
ggplot(mtcars, aes(x = mpg, y = hp)) +
  geom_point() +
  scale_x_continuous(breaks = pretty_breaks(n = 4))
}
\seealso{
\code{\link{format_num_as_nor_num_0}} is the default
  \code{formatting_fn}. \code{\link{every_nth}} thins labels on a discrete
  axis instead. The "Pretty breaks" section of \code{vignette("csstyle")}
  demonstrates \code{pretty_breaks()}.
}
