% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_palettes.R
\name{scale_color_cs}
\alias{scale_color_cs}
\title{ggplot2 color scale using Core Surveillance color palettes}
\usage{
scale_color_cs(..., palette = "primary", direction = 1)
}
\arguments{
\item{...}{Further arguments passed to \code{ggplot2::discrete_scale()}.}

\item{palette}{Name of the palette family. One of "primary", "warning" or
"posneg". Default "primary".}

\item{direction}{Direction of the palette. Use 1 for the normal order and -1
for the reversed order. Default 1.}
}
\value{
A ggplot2 discrete color scale.
}
\description{
Creates a discrete ggplot2 color scale from a predefined Core
  Surveillance color palette.
}
\examples{
library(ggplot2)

# Default primary color scale
ggplot(mtcars, aes(x = mpg, y = hp, color = factor(cyl))) +
  geom_point() +
  scale_color_cs()

# Use warning palette
ggplot(mtcars, aes(x = mpg, y = hp, color = factor(cyl))) +
  geom_point() +
  scale_color_cs(palette = "warning")
}
\seealso{
Other csstyle_ggplot2:
\code{\link[=scale_fill_cs]{scale_fill_cs()}},
\code{\link[=theme_cs]{theme_cs()}}
}
\concept{csstyle_ggplot2}
