% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.r
\name{view_html}
\alias{view_html}
\title{View HTML in the RStudio viewer}
\usage{
view_html(x)
}
\arguments{
\item{x}{Character string. The HTML content to show.}
}
\value{
Nothing (called for side effects).
}
\description{
Opens HTML content in the RStudio viewer pane.
}
\examples{
# View simple HTML content
html_content <- "<h1>Hello World</h1><p>This is a test.</p>"
\dontrun{
view_html(html_content)
}
}
\seealso{
\code{vignette("csstyle")} gives an overview of the package. It does
  not demonstrate \code{view_html()}.
}
