% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every_nth.R
\name{every_nth}
\alias{every_nth}
\title{Returns every nth discrete value}
\usage{
every_nth(n)
}
\arguments{
\item{n}{Integer specifying which nth value to return (e.g., n=2 returns every 2nd value)}
}
\value{
A function that returns every nth discrete value.
}
\description{
Creates a function that returns every nth element from a vector, useful for axis breaks in ggplot2.
}
\examples{
every_nth(4)(c(1:10))
\dontrun{
scale_x_discrete(NULL, breaks = every_nth(n = 2))
}
}
\seealso{
Other csstyle_utilities: 
\code{\link{colors}},
\code{\link{display_all_palettes}()},
\code{\link{set_global}()}
}
\concept{csstyle_utilities}
