% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num_as_nor_num_0}
\alias{format_num_as_nor_num_0}
\alias{format_num_as_nor_num_1}
\alias{format_num_as_nor_num_2}
\title{Format numbers with Norwegian conventions (0 digits)}
\usage{
format_num_as_nor_num_0(x)

format_num_as_nor_num_1(x)

format_num_as_nor_num_2(x)
}
\arguments{
\item{x}{Numeric vector to format}
}
\value{
Character vector with Norwegian-formatted numbers

Character vector with Norwegian-formatted numbers

Character vector with Norwegian-formatted numbers
}
\description{
Formats numeric values using Norwegian number conventions with 0 decimal places.

Formats numeric values using Norwegian number conventions with 1 decimal place.

Formats numeric values using Norwegian number conventions with 2 decimal places.
}
\examples{
# Basic number formatting
format_num_as_nor_num_0(c(1234.56, 9876.54, NA))
format_num_as_nor_num_1(c(1234.56, 9876.54, NA))
format_num_as_nor_num_2(c(1234.56, 9876.54, NA))
}
\seealso{
Other csstyle_formatters: 
\code{\link{format_num_as_nor_invlog10_1}()},
\code{\link{format_num_as_nor_invlog2_1}()},
\code{\link{format_num_as_nor_invloge_1}()},
\code{\link{format_num_as_nor_per100k_0}()},
\code{\link{format_num_as_nor_perc_0}()}

Other csstyle_formatters: 
\code{\link{format_num_as_nor_invlog10_1}()},
\code{\link{format_num_as_nor_invlog2_1}()},
\code{\link{format_num_as_nor_invloge_1}()},
\code{\link{format_num_as_nor_per100k_0}()},
\code{\link{format_num_as_nor_perc_0}()}

Other csstyle_formatters: 
\code{\link{format_num_as_nor_invlog10_1}()},
\code{\link{format_num_as_nor_invlog2_1}()},
\code{\link{format_num_as_nor_invloge_1}()},
\code{\link{format_num_as_nor_per100k_0}()},
\code{\link{format_num_as_nor_perc_0}()}
}
\concept{csstyle_formatters}
