% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_breaks.R
\name{pretty_breaks}
\alias{pretty_breaks}
\title{Pretty breaks}
\usage{
pretty_breaks(n = 5, formatting_fn = csstyle::format_num_as_nor_num_0, ...)
}
\arguments{
\item{n}{Number of desired breaks (default: 5)}

\item{formatting_fn}{Formatting function to apply to break labels (default: Norwegian number formatting)}

\item{...}{Additional arguments passed to pretty()}
}
\value{
A function that can be used for breaks on graphing.
}
\description{
Creates a function for generating pretty axis breaks with Norwegian number formatting.
}
\examples{
library(ggplot2)

# Use pretty breaks with default formatting
ggplot(mtcars, aes(x = mpg, y = hp)) +
  geom_point() +
  scale_x_continuous(breaks = pretty_breaks(n = 4))
}
\seealso{
\code{\link{format_num_as_nor_num_0}}, the default
  \code{formatting_fn}, and \code{\link{every_nth}}, which thins labels on a
  discrete axis instead. The "Pretty breaks" section of
  \code{vignette("csstyle")} demonstrates this function.
}
