% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_palettes.R
\name{scale_fill_cs}
\alias{scale_fill_cs}
\title{ggplot2 fill scale using Core Surveillance color palettes}
\usage{
scale_fill_cs(..., palette = "primary", direction = 1)
}
\arguments{
\item{...}{Further arguments passed to \code{ggplot2::discrete_scale()}.}

\item{palette}{Name of the palette family. One of "primary", "warning" or
"posneg". Default "primary".}

\item{direction}{Direction of the palette. Use 1 for the normal order and -1
for the reversed order. Default 1.}
}
\value{
A ggplot2 discrete fill scale.
}
\description{
Creates a discrete ggplot2 fill scale from a predefined Core
  Surveillance color palette.
}
\examples{
library(ggplot2)

# Default primary fill scale
ggplot(mtcars, aes(x = factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_cs()

# Use posneg palette
ggplot(mtcars, aes(x = factor(vs), fill = factor(vs))) +
  geom_bar() +
  scale_fill_cs(palette = "posneg")
}
\seealso{
Other csstyle_ggplot2:
\code{\link[=scale_color_cs]{scale_color_cs()}},
\code{\link[=theme_cs]{theme_cs()}}
}
\concept{csstyle_ggplot2}
