% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_html.r
\name{view_html}
\alias{view_html}
\title{View HTML in the RStudio viewer}
\usage{
view_html(x)
}
\arguments{
\item{x}{Character string containing HTML content to display}
}
\value{
Nothing (called for side effects).
}
\description{
Opens HTML content in the RStudio viewer pane.
}
\examples{
# View simple HTML content
html_content <- "<h1>Hello World</h1><p>This is a test.</p>"
\dontrun{
view_html(html_content)
}
}
\seealso{
\code{vignette("csstyle")} for an overview of the package.
  \code{view_html()} itself is not demonstrated there.
}
